Refactoring of Advanced Search
